feat(order_manager): plain-limit + standalone-stop entry path (ai-broker#39)
Replaces the bracket-order entry with a plain LimitOrderRequest followed
by a standalone StopOrderRequest attached on fill. Take-profit is managed
by the tick-loop's check_exits polling against target_price.

Why: Alpaca rejects fractional+bracket combos (code 42210000). PR #38's
whole-share floor in place_entry was a stopgap that dropped ~50% of
mean_reversion signals on the $1k live account by flooring SPY/QQQ/XLK
positions to 0 shares under the max-position-pct cap. Removing the floor
required all the latent fractional-truncation sites in the broader live
path to be fixed in lockstep, otherwise fractional positions would no-op
or silently leak.

Entry path (order_manager.py)
- place_entry: drop whole-share floor, submit plain LimitOrderRequest.
- _transition_to_open: on fill, submit standalone StopOrderRequest.
- Stop-attach failure: emergency_flatten + notify, flip DB to
  ENTRY_FAILED, pop from _active_orders, suppress trade_entry. Without
  this, the next stateless tick re-evaluates a ghost POSITION_OPEN row
  with no broker-side protection.
- EntryDecision.shares typed float (was int) — supports Alpaca's 1e-6
  fractional increment.
- Log format strings: %d → %.6f for fractional qty in
  emergency_flatten, place_trailing_stop.

Cascade fixes for fractional support across the live path
- main.check_exits: int(quantity) → float(...) so sub-1-share positions
  aren't silently skipped at exit time.
- main._wind_down_market: same fix; otherwise EOD wind-down leaves
  fractional intraday positions open overnight.
- recovery._reconcile: float qty + 1e-6 tolerance so fractional Alpaca
  vs DB drift isn't masked by int truncation collapsing both sides to 0.
- recovery._verify_stop_orders: float qty, otherwise emergency stop is
  refused for any sub-1-share Alpaca holding.
- recovery._eod_intraday_flatten: float qty, otherwise fractional
  intraday positions are skipped at EOD flatten — exposes the bot to
  the gap risk it was specifically designed to avoid.
- recovery._create_db_position / _close_db_position / _place_emergency_stop:
  float qty so Alpaca-discovered fractional positions land in the DB
  with their real qty (instead of 0) and side derivation works on the
  sign of the float.
- portfolio_assessor: float qty so cost_basis ≠ 0 for fractional
  holdings (P&L pct was always reporting 0% pre-fix).

Tests (test_order_manager_lifecycle.py)
- test_fill_attaches_standalone_stop: asserts plain StopOrderRequest is
  the second submit, alpaca_target_order_id stays None.
- test_stop_attach_failure_triggers_emergency_flatten: asserts
  emergency_flatten + notify + ENTRY_FAILED + pop + suppressed
  trade_entry — locks in the stop-attach race recovery.
- test_alpaca_42210000_simulated_no_rejection_under_new_path: stub
  Alpaca that rejects fractional+bracket but accepts plain limit;
  proves zero rejections under the new path (issue's named acceptance
  gate).
- test_fractional_fill_attaches_standalone_stop: end-to-end fractional
  entry → fill → fractional stop with the same qty.
- test_partial_fill_above_min_pct_accepted: updated to expect plain-stop
  submission instead of bracket-leg lookup.

Backtest A/B (13-ETF, 2024 full year, $1k cash, --no-regime-filter)
Branch and main produce byte-identical numbers (mean_reversion: 61
trades, +20.01%; overnight_drift: 744 trades, +20.93%) — backtest
doesn't traverse order_manager so the strategy/sizing surface is
unchanged. The win is that the live path now routes those exact
fractional trades to Alpaca instead of flooring them to 0.

Deferred (tracked in self_improve_followups.md)
- STOP_AND_TARGET_ACTIVE → STOP_ACTIVE rename + V11 migration. Status
  name retained for back-compat across 45 references in 10 files.
  Touching that pre-Monday is unjustified naming churn.

Acceptance gate the PR cannot satisfy
- Mon/Tue paper smoke test: real fractional entry → fill → standalone
  stop → exit. Runtime check, not a code change. Recommend running
  before flipping ALPACA_ENV=live.

Tests: 752 pass, ruff clean on edited files, mypy 1 fewer error than
main (no regression).

+++ trading_bot/strategy/portfolio_assessor.py
@@ -169,7 +169,11 @@ class PortfolioAssessor:
         market_value: float = float(position.get("market_value", 0.0))
         unrealized_pnl: float = float(position.get("unrealized_pnl", 0.0))
         avg_cost: float = float(position.get("avg_cost", 0.0))
-        quantity: int = int(position.get("quantity", 0))
+        # Float-typed: positions.quantity is fractional under the
+        # ai-broker#39 entry path. int() truncation of 0.43 → 0 would zero
+        # out cost_basis below and produce a 0% P&L for any sub-1-share
+        # holding.
+        quantity: float = float(position.get("quantity", 0))
 
         # Account is USD-only — Alpaca returns USD, no conversion needed
         value_usd: float = market_value
